Fixes for 64-bit compilation errors

+++ src/maidsafe/routing/routing_table.cc
@@ -210,7 +210,8 @@ bool RoutingTable::IsRemovable(const NodeId& node_id) {
   std::unique_lock<std::mutex> lock(mutex_);
   if (nodes_.size() < Parameters::routing_table_size_threshold)
     return false;
-  int sorted_count(PartialSortFromTarget(NodeId(), nodes_.size(), lock));
+  uint16_t sorted_count(PartialSortFromTarget(NodeId(), 
+      static_cast<uint16_t>(nodes_.size()), lock));
   if (sorted_count == 0)
     return false;
   size_t size(std::count_if(nodes_.begin(),
@@ -221,7 +222,7 @@ bool RoutingTable::IsRemovable(const NodeId& node_id) {
 if ((size < Parameters::split_avoidance) ||
     (size >= (Parameters::max_routing_table_size - Parameters::split_avoidance)))
     return false;
-  PartialSortFromTarget(node_id, nodes_.size(), lock);
+  PartialSortFromTarget(node_id, static_cast<uint16_t>(nodes_.size()), lock);
   if (nodes_[0].node_id != node_id)
     return false;
   if (NodeId::CloserToTarget(kNodeId(), nodes_[1].node_id, node_id))
@@ -418,7 +419,8 @@ NodeInfo RoutingTable::GetClosestNode(const NodeId& target_id,
 
 NodeInfo RoutingTable::GetClosestTo(const NodeId& node_id, bool backward) {
   std::unique_lock<std::mutex> lock(mutex_);
-  int sorted_count(PartialSortFromTarget(kNodeId(), nodes_.size(), lock));
+  uint16_t sorted_count(PartialSortFromTarget(kNodeId(), 
+      static_cast<uint16_t>(nodes_.size()), lock));
   if (sorted_count == 0)
     return NodeInfo();
   auto iterator(std::find_if(nodes_.begin(),
@@ -439,7 +441,8 @@ NodeInfo RoutingTable::GetClosestTo(const NodeId& node_id, bool backward) {
 NodeInfo RoutingTable::GetFurthestRemovableNode() {
   std::vector<NodeInfo> sorted_routing_table;
   std::unique_lock<std::mutex> lock(mutex_);
-  int sorted_count(PartialSortFromTarget(kNodeId(), nodes_.size(), lock));
+  uint16_t sorted_count(PartialSortFromTarget(kNodeId(), 
+      static_cast<uint16_t>(nodes_.size()), lock));
   if (sorted_count == 0)
     return NodeInfo();
   sorted_routing_table = nodes_;
